fuse: adjust linux specific logic
